+++ system/main.c
@@ -35,7 +35,7 @@ int 		Mask_matrix[BNST_MAX][BNST_MAX]; /* ÊÂÎó¥Þ¥¹¥¯
 double 		Para_matrix[PARA_MAX][BNST_MAX][BNST_MAX];
 double          Chi_case_prob_matrix[BNST_MAX][BNST_MAX];
 double          Chi_case_nominal_prob_matrix[BNST_MAX][BNST_MAX];
-int          Chi_gigaword_pa_matrix[BNST_MAX][BNST_MAX];  
+double          Chi_gigaword_pa_matrix[BNST_MAX][BNST_MAX];  
 int             Chi_np_start_matrix[BNST_MAX][BNST_MAX];
 int             Chi_np_end_matrix[BNST_MAX][BNST_MAX];
 int             Chi_quote_start_matrix[BNST_MAX][BNST_MAX];
